MainActivity.java
MainActivity.java file contains our main program written in JAVA. It has all the information regarding the function of our Application. In this file, I have written the program on the basis of the algorithm designed by myself. This file resides in what our Application is going to perform when it is integrated by the user. In other words what the buttons we designed in the activity_main.xml file are going to perform when they are clicked by the application user. I have added some safe codes to avoid the crashing or ill-performance of the program.